refactor: move style token files
