+++ src/client/components/atoms/Card/index.tsx
@@ -1,12 +1,12 @@
 /**@jsx jsx */
 import { css, jsx } from '@emotion/core';
 import * as React from 'react';
+import { SpacingKey } from 'client/styles/tokens/spacing';
+import { componentElevationKey } from 'client/styles/tokens/elevation';
+import { ThemeColorVariants } from 'client/styles/tokens/colors';
 import { useAppTheme } from 'client/styles/tokens';
-import { BorderRadiusKey } from 'client/styles/borderRadius';
-import { ThemeColorVariants } from 'client/styles/colors';
-import { SpacingKey } from 'client/styles/spacing';
+import { BorderRadiusKey } from 'client/styles/tokens/borderRadius';
 import { Surface, SurfaceProps } from 'client/components/atoms/Surface';
-import { componentElevationKey } from 'client/styles/elevation';
 import { BaseLink } from 'client/components/atoms/BaseLink';
 
 type CardContentProps = SurfaceProps & {
