+++ src/client/components/templates/Page/components/NavigationBar/index.tsx
@@ -10,7 +10,7 @@ import { Surface } from 'client/components/atoms/Surface';
 import { Typography } from 'client/components/atoms/Typography';
 import { useAppContext } from 'client/hooks/useAppContext';
 import { useOnClickOutside } from 'client/hooks/useOnClickOutside';
-import { componentElevationKey } from 'client/styles/elevation';
+import { componentElevationKey } from 'client/styles/tokens/elevation';
 import { commonStyles, useAppTheme } from 'client/styles/tokens';
 import {
   getDiscographyUrl,
