+++ src/client/features/Member/components/GlowStickBadge/index.tsx
@@ -1,7 +1,7 @@
 /**@jsx jsx */
 import { jsx, css } from '@emotion/core';
 import * as React from 'react';
-import { componentElevationKey } from 'client/styles/elevation';
+import { componentElevationKey } from 'client/styles/tokens/elevation';
 import { useAppTheme } from 'client/styles/tokens';
 import { GlowStickColorType } from 'server/actors/Members/constants/glowStickColor';
